example: consolidate to a single View/RecyclerView MainActivity
Remove the Compose demo screen and make the RecyclerView demo — the
integration shape publishers actually use (e.g. speakmaster) — the single
MainActivity and launcher, so there is one unambiguous entry point. Anyone
running `.MainActivity` (the conventional name) now gets the working View
demo. Renames activity_main_view.xml -> activity_main.xml; ChatMessage moves
into MainActivity.kt.
