+++ example/src/main/java/so/kontext/ads/example/MainActivity.kt
@@ -1,35 +1,25 @@
 package so.kontext.ads.example
 
 import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
 import android.util.Log
+import android.widget.Button
+import android.widget.EditText
+import android.widget.Switch
+import androidx.activity.ComponentActivity
+import androidx.lifecycle.lifecycleScope
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import android.view.View
-import androidx.compose.ui.viewinterop.AndroidView
 import so.kontext.ads.KontextAds
 import so.kontext.ads.Session
 import so.kontext.ads.model.AddMessageOptions
+import so.kontext.ads.model.AdEvent
 import so.kontext.ads.model.Message
 import so.kontext.ads.model.Role
 import so.kontext.ads.model.SessionOptions
-import so.kontext.ads.ui.InlineAdView
 
-private const val PLACEMENT_CODE = "inlineAd"
-
+/** A chat message rendered in the demo. */
 data class ChatMessage(
     val id: String,
     val role: Role,
@@ -39,325 +29,148 @@ data class ChatMessage(
     val showAd: Boolean = false,
 )
 
+/**
+ * Traditional-View (RecyclerView) demo — the integration shape publishers use
+ * from an XML / RecyclerView UI (e.g. the speakmaster app). Ads render through
+ * [so.kontext.ads.ui.InlineAdView], bound in [MessageAdapter]'s
+ * `onBindViewHolder`, with the RecyclerView scrolling/recycling natively.
+ */
 class MainActivity : ComponentActivity() {
+
+    private lateinit var session: Session
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContent {
-            MaterialTheme {
-                ChatScreen()
-            }
-        }
-    }
-}
+        setContentView(R.layout.activity_main)
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ChatScreen() {
-    val context = LocalContext.current
-    val scope = rememberCoroutineScope()
-    val listState = rememberLazyListState()
-
-    val session = remember {
-        // BuildConfig.AD_SERVER_URL is empty when the publisher hasn't set
-        // an override in local.properties — pass `null` in that case so the
-        // SDK falls back to its production default.
-        // (`Constants.DEFAULT_AD_SERVER_URL`).
         val adServerOverride: String? = BuildConfig.AD_SERVER_URL.takeIf { it.isNotBlank() }
-        KontextAds.createSession(
-            context = context.applicationContext,
+        session = KontextAds.createSession(
+            context = applicationContext,
             options = SessionOptions(
                 publisherToken = BuildConfig.PUBLISHER_TOKEN,
                 userId = "user-1",
                 conversationId = "conv-${System.currentTimeMillis()}",
                 adServerUrl = adServerOverride,
                 onEvent = { event -> Log.d("KontextExample", "Event: $event") },
-                // Drop the per-tick `update-dimensions-iframe` heartbeat
-                // (200 ms cadence) — useful for the SDK's viewability
-                // tracking but pure noise in the example's logcat. Every
-                // other SDK-internal diagnostic still flows through.
-                onDebugEvent = { event, data ->
-                    val typeStr = (data as? Map<*, *>)?.get("type") as? String
-                    if (typeStr != "update-dimensions-iframe") {
-                        Log.d("KontextExample", "Debug: $event $data")
+                onDebugEvent = { str, data ->
+                    if ((data as? Map<*, *>)?.get("type") != "update-dimensions-iframe") {
+                        Log.d("KontextExample", "Debug: $str $data")
                     }
                 },
             ),
         )
-    }
 
-    var messages by remember { mutableStateOf(listOf<ChatMessage>()) }
-    var input by remember { mutableStateOf("") }
-    var loading by remember { mutableStateOf(false) }
-    var lastAssistantId by remember { mutableStateOf<String?>(null) }
-    var msgCounter by remember { mutableIntStateOf(0) }
-    var trackOnly by remember { mutableStateOf(false) }
+        val recyclerView: RecyclerView = findViewById(R.id.messagesRecyclerView)
+        val input: EditText = findViewById(R.id.messageEditText)
+        val send: Button = findViewById(R.id.sendButton)
+        val trackOnlySwitch: Switch = findViewById(R.id.trackOnlySwitch)
 
-    DisposableEffect(session) {
-        onDispose { session.close() }
-    }
+        val adapter = MessageAdapter(session)
+        // Top-aligned list (messages start at the top), matching the customer's app.
+        recyclerView.layoutManager = LinearLayoutManager(this)
+        recyclerView.adapter = adapter
 
-    // Seed one user + one assistant message so the first /preload fires
-    // immediately on launch (rather than waiting for the publisher to
-    // type something). Mirrors sdk-swift's seedDemoMessages.
-    LaunchedEffect(Unit) {
-        val seedUser = ChatMessage("u-seed", Role.USER, "Hi Aria!")
-        val seedAssistant = ChatMessage(
-            "a-seed",
-            Role.ASSISTANT,
-            "Hi! I'm Aria, your friendly AI companion.",
-        )
-        messages = listOf(seedUser, seedAssistant)
-        lastAssistantId = seedAssistant.id
-        session.addMessage(Message(id = seedUser.id, role = Role.USER, content = seedUser.content))
-        session.addMessage(
-            Message(id = seedAssistant.id, role = Role.ASSISTANT, content = seedAssistant.content),
-        )
-    }
+        val messages = mutableListOf<ChatMessage>()
+        var counter = 0
+        // True between sending a user message and the assistant reply arriving.
+        // While loading, no ad is shown — so the previous ad disappears the
+        // instant the user sends, and the new one appears with the reply.
+        var loading = false
 
-    LaunchedEffect(messages.size) {
-        if (messages.isNotEmpty()) {
-            listState.animateScrollToItem(messages.size - 1)
+        // Reveal the bottom edge of the last item (the ad sits at the bottom of
+        // its message and may exceed the viewport). Only ever scrolls DOWN, so
+        // it can't yank a user who scrolled up.
+        fun pinLastItemBottom() {
+            val lm = recyclerView.layoutManager as LinearLayoutManager
+            val last = adapter.itemCount - 1
+            if (last < 0) return
+            val view = lm.findViewByPosition(last) ?: return
+            val overflow = view.bottom - (recyclerView.height - recyclerView.paddingBottom)
+            if (overflow > 0) recyclerView.scrollBy(0, overflow)
         }
-    }
 
-    // When an ad becomes visible after the assistant message is already
-    // on screen, scroll back to the bottom — BUT only when the user is
-    // already pinned to the bottom. A reader who scrolled up to re-read
-    // an earlier message shouldn't get yanked back down by an ad-grow
-    // event. Listens to: Filled (bid resolved), AdHeight (height grew),
-    // RenderCompleted (iframe fully laid out).
-    LaunchedEffect(session) {
-        session.events.collect { event ->
-            when (event) {
-                is so.kontext.ads.model.AdEvent.Filled,
-                is so.kontext.ads.model.AdEvent.AdHeight,
-                is so.kontext.ads.model.AdEvent.RenderCompleted,
-                -> {
-                    if (messages.isNotEmpty() && listState.isAtBottom()) {
-                        listState.animateScrollToItem(messages.size - 1)
-                    }
-                }
-                else -> Unit
-            }
+        fun scrollToBottom() {
+            val last = adapter.itemCount - 1
+            if (last < 0) return
+            recyclerView.scrollToPosition(last)
+            recyclerView.post { pinLastItemBottom() }
         }
-    }
 
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text("Kontext v4 — Kotlin") },
-                actions = {
-                    Text(
-                        text = "Track only",
-                        style = MaterialTheme.typography.labelMedium,
-                        modifier = Modifier.padding(end = 4.dp),
-                    )
-                    // Mirrors sdk-swift's Track-only switch on the example's
-                    // nav bar: user messages route through
-                    // AddMessageOptions(trackOnly = true), which marks the
-                    // resulting /preload as analytics-only (no bid resolution).
-                    Switch(
-                        checked = trackOnly,
-                        onCheckedChange = { trackOnly = it },
-                        modifier = Modifier.padding(end = 8.dp),
-                    )
-                },
-            )
-        },
-    ) { padding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(padding)
-                // Pulls the input bar above the IME so a long thread doesn't
-                // hide the latest message behind the keyboard. The system-bar
-                // inset is already handled by Scaffold's padding above.
-                .imePadding(),
-        ) {
-            LazyColumn(
-                state = listState,
-                modifier = Modifier
-                    .weight(1f)
-                    .fillMaxWidth()
-                    .padding(horizontal = 12.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-                contentPadding = PaddingValues(vertical = 12.dp),
-            ) {
-                items(messages, key = { it.id }) { msg ->
-                    // Column groups the bubble + its (optional) ad so they
-                    // share one LazyColumn slot. The 8 dp Spacer below is
-                    // only emitted when the ad actually renders, so we get
-                    // breathing room between the assistant bubble and the
-                    // ad without adding dead space to other rows.
-                    Column {
-                        MessageBubble(msg)
-                        if (msg.role == Role.ASSISTANT && msg.id == lastAssistantId && !loading) {
-                            Spacer(modifier = Modifier.height(8.dp))
-                            // The example renders ads through the traditional-View
-                            // `InlineAdView` interop wrapper — matching publishers
-                            // integrating from a RecyclerView/XML UI (e.g. speakmaster).
-                            InlineAdViewHost(messageId = msg.id, session = session)
-                        }
-                    }
-                }
+        fun isNearBottom(): Boolean {
+            val lm = recyclerView.layoutManager as LinearLayoutManager
+            return lm.findLastVisibleItemPosition() >= adapter.itemCount - 2
+        }
 
-                if (loading) {
-                    item {
-                        Text(
-                            text = "Thinking...",
-                            style = MaterialTheme.typography.bodySmall,
-                            modifier = Modifier.padding(start = 8.dp, top = 4.dp),
-                        )
-                    }
-                }
+        fun publish() {
+            // Only the latest assistant message shows the ad (and only when not
+            // loading) — older ones flip to showAd=false, so DiffUtil rebinds
+            // them and clears their ad.
+            val lastAssistantId = messages.lastOrNull { it.role == Role.ASSISTANT }?.id
+            val display = messages.map {
+                it.copy(showAd = !loading && it.role == Role.ASSISTANT && it.id == lastAssistantId)
             }
+            adapter.submitList(display) { scrollToBottom() }
+        }
 
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(12.dp),
-                verticalAlignment = Alignment.Bottom,
-            ) {
-                OutlinedTextField(
-                    value = input,
-                    onValueChange = { input = it },
-                    modifier = Modifier.weight(1f),
-                    placeholder = { Text("Type a message...") },
-                    minLines = 1,
-                    maxLines = 4,
-                    enabled = !loading,
+        fun sendUserMessage(text: String, trackOnly: Boolean) {
+            counter++
+            val userId = "u$counter"
+            messages.add(ChatMessage(userId, Role.USER, text))
+            loading = true
+            publish()
+            lifecycleScope.launch {
+                // Only user messages carry trackOnly — they're what triggers
+                // the /preload. trackOnly => analytics-only, no bid resolution.
+                session.addMessage(
+                    Message(id = userId, role = Role.USER, content = text),
+                    AddMessageOptions(trackOnly = trackOnly),
                 )
+                delay(500)
+                counter++
+                val assistantId = "a$counter"
+                val reply = "This is a response from the assistant."
+                messages.add(ChatMessage(assistantId, Role.ASSISTANT, reply))
+                loading = false
+                publish()
+                session.addMessage(Message(id = assistantId, role = Role.ASSISTANT, content = reply))
+            }
+        }
 
-                Spacer(modifier = Modifier.width(8.dp))
+        send.setOnClickListener {
+            val text = input.text.toString().trim()
+            if (text.isNotEmpty()) {
+                input.text.clear()
+                sendUserMessage(text, trackOnlySwitch.isChecked)
+            }
+        }
 
-                Button(
-                    onClick = {
-                        val trimmed = input.trim()
-                        if (trimmed.isEmpty() || loading) return@Button
-
-                        loading = true
-                        input = ""
-                        msgCounter++
-                        val userId = "u$msgCounter"
-
-                        messages = messages + ChatMessage(userId, Role.USER, trimmed)
-
-                        // Snapshot the trackOnly switch state at send time
-                        // so a mid-flight toggle doesn't apply retroactively
-                        // to the assistant message handler.
-                        val trackOnlyAtSendTime = trackOnly
-
-                        scope.launch {
-                            // Fire-and-forget: preload runs in background
-                            launch {
-                                session.addMessage(
-                                    Message(id = userId, role = Role.USER, content = trimmed),
-                                    AddMessageOptions(trackOnly = trackOnlyAtSendTime),
-                                )
-                            }
-
-                            // Simulate assistant response
-                            delay(500)
-                            msgCounter++
-                            val assistantId = "a$msgCounter"
-                            val assistantContent = "This is a response from the assistant."
-
-                            messages = messages + ChatMessage(assistantId, Role.ASSISTANT, assistantContent)
-                            lastAssistantId = assistantId
-
-                            // Fire-and-forget: notify SDK of assistant message.
-                            // Assistant messages don't carry trackOnly — only
-                            // user messages trigger preloads, so flag-on-user
-                            // is the right scope (matches sdk-swift).
-                            launch {
-                                session.addMessage(
-                                    Message(id = assistantId, role = Role.ASSISTANT, content = assistantContent),
-                                )
-                            }
-
-                            loading = false
-                        }
-                    },
-                    enabled = input.trim().isNotEmpty() && !loading,
-                ) {
-                    Text("Send")
+        // The ad fills/resizes asynchronously after its message is added, so
+        // reveal its bottom when the height changes — but only if the user is
+        // already near the bottom (don't yank them up if they scrolled to read).
+        lifecycleScope.launch {
+            session.events.collect { event ->
+                when (event) {
+                    is AdEvent.Filled, is AdEvent.AdHeight, is AdEvent.RenderCompleted -> {
+                        if (isNearBottom()) recyclerView.post { pinLastItemBottom() }
+                    }
+                    else -> Unit
                 }
             }
+        }
+
+        // Seed an initial greeting (mirrors sdk-swift's seedDemoMessages) so
+        // the first /preload fires on launch and the reply shows an ad.
+        lifecycleScope.launch {
+            messages.add(ChatMessage("u-seed", Role.USER, "Hi Aria!"))
+            messages.add(ChatMessage("a-seed", Role.ASSISTANT, "Hi! I'm Aria, your friendly AI companion."))
+            publish()
+            session.addMessage(Message(id = "u-seed", role = Role.USER, content = "Hi Aria!"))
+            session.addMessage(Message(id = "a-seed", role = Role.ASSISTANT, content = "Hi! I'm Aria, your friendly AI companion."))
         }
     }
-}
 
-/**
- * Whether the LazyColumn is currently pinned at (or within one item of)
- * the bottom. Read at event-collect time to gate the auto-scroll on ad
- * height changes — keeps a user who scrolled up from being yanked
- * back down.
- */
-private fun LazyListState.isAtBottom(): Boolean {
-    val info = layoutInfo
-    val lastVisible = info.visibleItemsInfo.lastOrNull()?.index ?: return true
-    return lastVisible >= info.totalItemsCount - 2
-}
-
-/**
- * Hosts the traditional-View [InlineAdView] inside Compose via [AndroidView],
- * mirroring how a RecyclerView/XML publisher integrates (e.g. speakmaster):
- * construct the view, set `onHeightChange`, and call `bind(messageId, session)`
- * once the view is attached to the window — matching a RecyclerView's
- * `onBindViewHolder`, where `itemView` is already attached so
- * `findViewTreeLifecycleOwner()` resolves.
- */
-@Composable
-fun InlineAdViewHost(
-    messageId: String,
-    session: Session,
-    modifier: Modifier = Modifier,
-) {
-    AndroidView(
-        modifier = modifier.fillMaxWidth(),
-        factory = { ctx ->
-            InlineAdView(ctx).apply {
-                onHeightChange = { h ->
-                    Log.d("KontextExample", "InlineAdView onHeightChange: $h")
-                }
-                addOnAttachStateChangeListener(object : View.OnAttachStateChangeListener {
-                    override fun onViewAttachedToWindow(v: View) {
-                        bind(messageId, session)
-                    }
-
-                    override fun onViewDetachedFromWindow(v: View) = Unit
-                })
-            }
-        },
-    )
-}
-
-@Composable
-fun MessageBubble(msg: ChatMessage) {
-    val isUser = msg.role == Role.USER
-
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = if (isUser) Arrangement.End else Arrangement.Start,
-    ) {
-        Surface(
-            shape = RoundedCornerShape(
-                topStart = 16.dp,
-                topEnd = 16.dp,
-                bottomStart = if (isUser) 16.dp else 4.dp,
-                bottomEnd = if (isUser) 4.dp else 16.dp,
-            ),
-            color = if (isUser)
-                MaterialTheme.colorScheme.primaryContainer
-            else
-                MaterialTheme.colorScheme.surfaceVariant,
-            modifier = Modifier.widthIn(max = 280.dp),
-        ) {
-            Text(
-                text = msg.content,
-                modifier = Modifier.padding(12.dp),
-                style = MaterialTheme.typography.bodyMedium,
-            )
-        }
+    override fun onDestroy() {
+        super.onDestroy()
+        session.close()
     }
 }
